Upgrade ChatApp to 1.0.0-alpha10
