+++ ChatApp/app/src/main/AndroidManifest.xml
@@ -15,7 +15,8 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:appfn="http://schemas.android.com/apk/androidx.appfunctions">
+    xmlns:appfn="http://schemas.android.com/apk/androidx.appfunctions"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <application
         android:name=".ChatApplication"
@@ -49,6 +50,21 @@
                     android:host="com.example.chatapp" />
             </intent-filter>
         </activity>
+        <service
+            android:name="com.example.chatapp.appfunctions.ChatAppFunctionService"
+            android:permission="android.permission.BIND_APP_FUNCTION_SERVICE"
+            android:exported="true"
+            tools:targetApi="36">
+            <property
+                android:name="android.app.appfunctions.schema"
+                android:value="app_functions_schema.xsd" />
+            <property
+                android:name="android.app.appfunctions.v2"
+                android:value="chat_app_function_service.xml" />
+            <intent-filter>
+                <action android:name="android.app.appfunctions.AppFunctionService" />
+            </intent-filter>
+        </service>
         <property
             android:name="android.app.appfunctions.app_metadata"
             android:resource="@xml/app_metadata" />
