+++ ChatApp/app/src/main/res/xml/app_metadata.xml
@@ -14,13 +14,13 @@
   ~ limitations under the License.
   -->
 
-<AppFunctionAppMetadata xmlns:appfn="http://schemas.android.com/apk/res-auto"
-    appfn:description="Provides tools for managing chat conversations, searching contacts, and initiating calls.
+<AppFunctionAppMetadata xmlns:appfn="http://schemas.android.com/apk/androidx.appfunctions"
+    appfn:description="Provides functions for managing chat conversations, searching contacts, sending messages, and initiating voice calls.
 Operational Patterns:
-  - Resolution of contacts via 'searchContacts' provides the required 'endpointValue' for sending messages or making calls.
-  - Passing a blank query to 'searchContacts' retrieves the most recent contacts.
+  - Always call 'searchContacts' first to resolve contact or group names to an 'endpointValue' before calling 'send' or 'makeCall'.
+  - Passing a blank query string to 'searchContacts' retrieves the most recently contacted entities.
 Constraints:
-  - Message content cannot be empty or blank.
-  - Calls are initiated by unique identifier.
-  - Media attachments are supported via a list of URIs."
+  - Message content cannot be empty or blank when calling 'send'.
+  - Calls and messages must use valid endpointValue identifiers returned by 'searchContacts'.
+  - Image attachments are supported in 'send' via a list of content URIs."
     appfn:displayDescription="@string/app_function_app_display_description"/>
