+++ ChatApp/shared/src/main/kotlin/com/example/chatapp/appfunctions/AppFunctions.kt
@@ -15,195 +15,63 @@
  */
 package com.example.chatapp.appfunctions
 
-import android.app.PendingIntent
-import android.content.Intent
-import android.net.Uri
-import androidx.appfunctions.AppFunctionAppUnknownException
-import androidx.appfunctions.AppFunctionContext
-import androidx.appfunctions.AppFunctionElementNotFoundException
-import androidx.appfunctions.AppFunctionInvalidArgumentException
 import androidx.appfunctions.AppFunctionSerializable
-import androidx.appfunctions.AppFunctionStringValueConstraint
-import androidx.appfunctions.service.AppFunction
-import com.example.chatapp.data.CallManager
-import com.example.chatapp.data.MessageRepository
-import com.example.chatapp.data.RecipientsRepository
-import javax.inject.Inject
 
 /**
- * Provides functions for chat-related operations such as retrieving contacts and sending messages.
+ * Identifiers for AppFunctions exposed by ChatApp.
  */
-class AppFunctions
-    @Inject
-    constructor(
-        private val messageRepository: MessageRepository,
-        private val recipientsRepository: RecipientsRepository,
-        private val callManager: CallManager,
-    ) {
-        /**
-         * Search for contacts or groups by name.
-         *
-         * @param appFunctionContext The context of this app function call.
-         * @param query Search string for the contact or group name. Can be partial or full names.
-         *   If blank, returns the most recently contacted entities.
-         * @param filterType Filter results by entity type. Accepts "INDIVIDUAL" or "GROUP".
-         */
-        @AppFunction(isDescribedByKDoc = true)
-        suspend fun searchContacts(
-            appFunctionContext: AppFunctionContext,
-            query: String,
-            @AppFunctionStringValueConstraint(enumValues = ["INDIVIDUAL", "GROUP"])
-            filterType: String,
-        ): List<ContactSearchResult> {
-            val recipients =
-                when (filterType) {
-                    "INDIVIDUAL" -> {
-                        recipientsRepository.searchRecipients(query, 3).map {
-                            ContactSearchResult(
-                                endpointValue = it.id,
-                                endpointType = "INDIVIDUAL",
-                                displayName = it.name,
-                            )
-                        }
-                    }
-                    "GROUP" -> {
-                        recipientsRepository.searchGroups(query, 3).map {
-                            ContactSearchResult(
-                                endpointValue = it.id,
-                                endpointType = "GROUP",
-                                displayName = it.name,
-                            )
-                        }
-                    }
-                    else -> {
-                        recipientsRepository.searchAny(query, maxCount = 3)
-                            .ifEmpty {
-                                throw AppFunctionInvalidArgumentException(
-                                    "Only INDIVIDUAL or GROUP are accepted filter arguments.",
-                                )
-                            }
-                    }
-                }
+object AppFunctionsIds {
+    const val SEND_ID = "com.example.chatapp.appfunctions.BaseChatAppFunctionService#send"
+    const val SEARCH_CONTACTS_ID = "com.example.chatapp.appfunctions.BaseChatAppFunctionService#searchContacts"
+    const val MAKE_CALL_ID = "com.example.chatapp.appfunctions.BaseChatAppFunctionService#makeCall"
+}
 
-            if (recipients.isEmpty()) {
-                throw AppFunctionInvalidArgumentException(
-                    "$filterType with name $query not found. Ask the user to clarify the name",
-                )
-            }
-            return recipients
-        }
+/**
+ * Represents a result from a contact or group search.
+ */
+@AppFunctionSerializable(isDescribedByKDoc = true)
+data class ContactSearchResult(
+    /** The unique identifier for the contact or group. */
+    val endpointValue: String,
+    /** The type of the found entity, either "INDIVIDUAL" or "GROUP". */
+    val endpointType: String,
+    /** The human-readable display name of the contact or group. */
+    val displayName: String,
+)
 
-        /**
-         * Send a text message with optional image attachments.
-         *
-         * @param appFunctionContext The context of this app function call.
-         * @param endpointValue The unique identifier for the recipient or group.
-         * @param messageBody The text content of the message. Cannot be empty.
-         * @param imageUris List of URIs for images to attach.
-         */
-        @AppFunction(isDescribedByKDoc = true)
-        suspend fun send(
-            appFunctionContext: AppFunctionContext,
-            endpointValue: String,
-            messageBody: String,
-            imageUris: List<Uri>? = null,
-        ): Result {
-            if (messageBody.isBlank()) {
-                throw AppFunctionInvalidArgumentException("Message body cannot be empty")
-            }
-            val displayName =
-                recipientsRepository.getRecipientById(endpointValue)?.name
-                    ?: recipientsRepository.getGroupById(endpointValue)?.name
-                    ?: throw AppFunctionElementNotFoundException(
-                        "No contact or group found for endpointValue: $endpointValue",
-                    )
+/**
+ * Result of a message sending operation.
+ */
+@AppFunctionSerializable(isDescribedByKDoc = true)
+data class Result(
+    /** The unique identifier for the successfully sent message. */
+    val messageId: String,
+    /** A human-readable status message confirming success. */
+    val message: String,
+)
 
-            val sentMessageId =
-                try {
-                    messageRepository.send(
-                        text = messageBody,
-                        recipientIds = listOf(endpointValue),
-                        imageUris = imageUris?.map { it.toString() },
-                    )
-                } catch (e: Exception) {
-                    throw AppFunctionAppUnknownException("Failed to send message: ${e.message}")
-                }
+/**
+ * Represents an individual recipient or contact.
+ */
+@AppFunctionSerializable(isDescribedByKDoc = true)
+data class Recipient(
+    /** Unique identifier for the contact. */
+    val id: String,
+    /** Human-readable name of the contact. */
+    val name: String,
+    /** Email address of the contact. */
+    val email: String,
+)
 
-            // 3. RETURN: Provide a confirmation string
-            // The bot may use this string or the fact that it didn't throw to confirm success.
-            return Result(sentMessageId, "Message sent to: $displayName.")
-        }
-
-        /**
-         * Initiate a voice call.
-         *
-         * @param appFunctionContext The context of this app function call.
-         * @param endpointValue The unique identifier for the recipient.
-         */
-        @AppFunction(isDescribedByKDoc = true)
-        suspend fun makeCall(
-            appFunctionContext: AppFunctionContext,
-            endpointValue: String,
-        ): PendingIntent {
-            val recipient =
-                recipientsRepository.getRecipientById(endpointValue)
-                    ?: throw AppFunctionElementNotFoundException(
-                        "No recipient exists for endpointValue: $endpointValue",
-                    )
-
-            // Call manager should technically also record it here depending on app architecture,
-            // but we will launch the intent to handle it in the UI.
-            callManager.startCall(recipient.id)
-
-            // Create a pending intent to the MainActivity, initiating the call
-            return PendingIntent.getActivity(
-                appFunctionContext.context,
-                0,
-                Intent(appFunctionContext.context, Class.forName("com.example.chatapp.MainActivity"))
-                    .apply { putExtra("nav_route", "call/${recipient.id}") },
-                PendingIntent.FLAG_ONE_SHOT or PendingIntent.FLAG_IMMUTABLE,
-            )
-        }
-
-        /** Represents a result from a contact or group search. */
-        @AppFunctionSerializable(isDescribedByKDoc = true)
-        data class ContactSearchResult(
-            /** The unique identifier. */
-            val endpointValue: String,
-            /** The type of the found entity, either "INDIVIDUAL" or "GROUP". */
-            val endpointType: String,
-            /** The human-readable name of the contact or group. */
-            val displayName: String,
-        )
-
-        /** Result of a message sending operation. */
-        @AppFunctionSerializable(isDescribedByKDoc = true)
-        data class Result(
-            /** The unique identifier for the successfully sent message. */
-            val messageId: String,
-            /** A human-readable status message indicating success. */
-            val message: String,
-        )
-
-        /** Represents a recipient of a message. */
-        @AppFunctionSerializable(isDescribedByKDoc = true)
-        data class Recipient(
-            /** Unique identifier for the contact. */
-            val id: String,
-            /** Human-readable name. */
-            val name: String,
-            /** Email address of the contact. */
-            val email: String,
-        )
-
-        /** Represents a group of recipients. */
-        @AppFunctionSerializable(isDescribedByKDoc = true)
-        data class ChatGroup(
-            /** Unique identifier for the group. */
-            val id: String,
-            /** Name of the group. */
-            val name: String,
-            /** List of members belonging to the group. */
-            val recipients: List<Recipient>,
-        )
-    }
+/**
+ * Represents a group of chat recipients.
+ */
+@AppFunctionSerializable(isDescribedByKDoc = true)
+data class ChatGroup(
+    /** Unique identifier for the group. */
+    val id: String,
+    /** Name of the group. */
+    val name: String,
+    /** List of members belonging to the group. */
+    val recipients: List<Recipient>,
+)
