+++ ChatApp/wear/build.gradle.kts
@@ -28,7 +28,7 @@ android {
     defaultConfig {
         applicationId = "com.example.chatapp.wear"
         minSdk = 30
-        targetSdk = 35
+        targetSdk = 37
         versionCode = 1
         versionName = "1.0"
     }
@@ -66,7 +66,6 @@ dependencies {
 
     // App functions
     implementation(libs.androidx.appfunctions)
-    implementation(libs.androidx.appfunctions.service)
     ksp(libs.androidx.appfunctions.compiler)
 
 }
