+++ ChatApp/wear/src/main/res/xml/app_metadata.xml
@@ -15,6 +15,12 @@
   ~ limitations under the License.
   -->
 
-<AppFunctionAppMetadata xmlns:appfn="http://schemas.android.com/apk/res-auto"
-    appfn:description="Provides tools for managing chat conversations on WearOS."
+<AppFunctionAppMetadata xmlns:appfn="http://schemas.android.com/apk/androidx.appfunctions"
+    appfn:description="Provides functions for managing chat conversations on WearOS, searching contacts, sending messages, and initiating voice calls.
+Operational Patterns:
+  - Always call 'searchContacts' first to resolve contact or group names to an 'endpointValue' before calling 'send' or 'makeCall'.
+  - Passing a blank query string to 'searchContacts' retrieves the most recently contacted entities.
+Constraints:
+  - Message content cannot be empty or blank when calling 'send'.
+  - Calls and messages must use valid endpointValue identifiers returned by 'searchContacts'."
     appfn:displayDescription="@string/app_function_app_display_description"/>
